Move create funnel helpers to lib
